blogproject/requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
black==25.1.0
click==8.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
flake8==7.3.0
isort==6.0.1
Markdown==3.8.2
mccabe==0.7.0
mypy==1.17.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
soupsieve==2.7
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.1
wheel==0.45.1
